requirements-spanner.txt
# Google Spanner dependencies for TextQL MCP
google-auth
google-cloud-spanner
